+++ src/lib.rs
@@ -102,9 +102,10 @@ pub use udf::{
     PartitionDescriptor, ShortAd, TagId, UdfFile, UdfVolume,
 };
 pub use vob::{
-    demux_vobs, demux_vobs_path, looks_like_nav_pack, ButtonInfo, CellId, DsiPacket, DvdSubstream,
-    ElementaryStream, HighlightInfo, NavPack, PackHeader, PciPacket, PesPacket, SlColi, SlColiCell,
-    VobDemuxer, VobId, VobStreams,
+    demux_vobs, demux_vobs_path, looks_like_nav_pack, ButtonInfo, CellId, DsiGi, DsiPacket,
+    DvdSubstream, ElementaryStream, HighlightInfo, NavPack, PackHeader, PciPacket, PesPacket,
+    SlColi, SlColiCell, SmlAgli, SmlAngleCell, SmlAudioGap, SmlPbi, Synci, VobDemuxer, VobId,
+    VobStreams, VobuSri,
 };
 
 #[cfg(feature = "registry")]